digital-signage-saas/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate>=4.0.7
Werkzeug==3.0.1
stripe==7.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
requests>=2.31.0
PyJWT>=2.8.0
